Login finished add query delete  detials finished

+++ src/main/java/tk/ANPRCloud/service/NumberPlateManagerImpl.java
@@ -1,6 +1,7 @@
 package tk.ANPRCloud.service;
 
 import java.io.UnsupportedEncodingException;
+import java.util.ArrayList;
 import java.util.List;
 
 import org.apache.log4j.Logger;
@@ -29,7 +30,7 @@ public class NumberPlateManagerImpl implements NumberPlateManager {
 	//This method will be called when a numberPlate object is added
 	@Override
 	@Transactional
-	public int addNumberPlate(String username, NumberPlateFile numberPlateFile) {
+	public List<Object> addNumberPlate(String username, NumberPlateFile numberPlateFile) {
 		numberPlateEntity = new NumberPlateEntity();
 		// Set the numberPlateEntity properties
 		numberPlateProcessing = new NumberPlateProcessing(numberPlateFile.getImage());
@@ -39,7 +40,10 @@ public class NumberPlateManagerImpl implements NumberPlateManager {
 		numberPlateEntity.setOwner(username);
 		numberPlateEntity.setNumber(numberPlateProcessing.calculateNumber());
 		numberPlateEntity.setDetails(numberPlateProcessing.getDetails());
-		return numberPlateDAO.addNumberPlate(numberPlateEntity);
+		List<Object> idAndNumber = new ArrayList<Object>();
+		idAndNumber.add(numberPlateDAO.addNumberPlate(numberPlateEntity));
+		idAndNumber.add(numberPlateEntity.getNumber());
+		return idAndNumber;
 	}
 	
 	//This method return list of numberPlates in database
@@ -70,14 +74,32 @@ public class NumberPlateManagerImpl implements NumberPlateManager {
 		}
 	}
 	
-	//Get details of numberPlate by it's id
+	//Get image of numberPlate by it's id
 	@Override
 	@Transactional
 	public NumberPlateEntity queryNumberPlate(String username, Integer numberPlateId) throws InvalidFrontEndAccessException {
 		try {
 			if (numberPlateDAO.authorizeUsernameAndId(username, numberPlateId)){
 				NumberPlateEntity entity = numberPlateDAO.queryNumberPlate(numberPlateId);
-				numberPlateDAO.deleteNumberPlate(entity.getId());
+				//if (username.equals("")) {numberPlateDAO.deleteNumberPlate(entity.getId());}
+				return entity;
+			} else {
+				throw new InvalidFrontEndAccessException();
+			}
+		} catch (InvalidDataAccessException e){
+			logger.info("No any number plate or data access error!");
+			return null;
+		}
+	}
+	
+	//Get image of numberPlate by it's id
+	@Override
+	@Transactional
+	public NumberPlateEntity queryNumberPlateDetails(String username, Integer numberPlateId) throws InvalidFrontEndAccessException {
+		try {
+			if (numberPlateDAO.authorizeUsernameAndId(username, numberPlateId)){
+				NumberPlateEntity entity = numberPlateDAO.queryNumberPlateDetails(numberPlateId);
+				if (username.equals("")) {numberPlateDAO.deleteNumberPlate(entity.getId());}
 				return entity;
 			} else {
 				throw new InvalidFrontEndAccessException();
